Move login to Paywalls screen
