+++ Tests/TestingApps/PaywallsTester/PaywallsTester/SimpleApp.swift
@@ -12,7 +12,7 @@ struct SimpleApp: App {
 
     var body: some Scene {
         WindowGroup {
-            RootScreen()
+            AppContentView()
         }
     }
